requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argh==0.26.2
astor==0.8.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
backcall==0.2.0
backports.cached-property==1.0.2
beautifulsoup4==4.12.2
cachetools==5.3.1
Cerberus==1.3.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==1.3.0
colorama==0.4.6
conda-flow==0.3.3
configparser==4.0.2
cycler==0.10.0
Cython==0.29.35
dcargs==0.1.10
decorator==4.4.1
dm-control==1.0.12
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring-parser==0.15
fasteners==0.18
filelock==3.12.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
gast==0.3.2
gdown==4.7.1
geomloss==0.2.6
gitdb==4.0.10
gitdb2==2.0.6
GitPython==3.1.31
glfw==2.5.9
google==3.0.0
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.1.8
gql==0.1.0
graphql-core==2.2.1
grpcio==1.54.2
gym==0.17.1
gym-minigrid==1.0.3
h5py==2.10.0
higher==0.2.1
hydra-core==1.3.2
idna==2.8
imageio==2.4.1
imageio-ffmpeg==0.3.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
ipdb==0.13.13
ipython==7.34.0
jedi==0.19.0
joblib==1.2.0
jsonschema==4.17.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
labmaze==1.0.6
libtmux==0.8.2
lockfile==0.12.2
lxml==4.9.2
Markdown==3.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
moviepy==0.2.3.5
msgpack==0.6.2
mujoco==2.3.5
mujoco-py==2.1.2.14
multidict==6.0.4
networkx==2.6.3
numpy===1.17.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py3==7.352.0
oauthlib==3.0.0
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pandas==1.3.5
parso==0.8.3
patchelf==0.17.2.1
pathfinding==0.0.4
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
pkgutil_resolve_name==1.3.10
proglog==0.1.9
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.6.7
ptyprocess==0.7.0
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pycparser==2.19
pyglet==1.4.8
Pygments==2.16.1
PyOpenGL==3.1.7
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2023.3.post1
PyVirtualDisplay==3.0
PyWavelets==1.3.0
PyYAML==5.2
ray
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.1
rsa==4.9
Rx==1.6.1
scikit-image==0.18.3
scikit-learn==1.0.2
scikit-learn
scipy==1.4.0
seaborn==0.10.1
sentry-sdk==1.9.0
setproctitle==1.3.2
shortuuid==0.5.0
six==1.13.0
smmap==5.0.0
smmap2==2.0.5
soupsieve==2.4.1
subprocess32==3.5.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.40.2
traitlets==5.9.0
typing==3.7.4.3
typing_extensions==4.6.3
urllib3==1.25.7
wandb==0.16.5
watchdog==0.9.0
wcwidth==0.2.6
Werkzeug==0.16.0
wrapt==1.11.2
yarl==1.9.2
zipp==3.15.0
